// app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  MANAGER_MULTI: 'MANAGER_MULTI',
  CAISSIER: 'CAISSIER',
  CAISSE_RESTAURANT: 'CAISSE_RESTAURANT',
  CAISSE_BAR: 'CAISSE_BAR',
  ECONOMAT: 'ECONOMAT',
  MAGASINIER: 'MAGASINIER',
  LOCATION: 'LOCATION',
  CAISSE_PHARMACIE: 'CAISSE_PHARMACIE',
  PHARMACIE: 'PHARMACIE',
  STOCK: 'STOCK',
  OTHER: 'OTHER',
  CONSEIL_ADMINISTRATION: 'CONSEIL_ADMINISTRATION',
  SUPERVISEUR: 'SUPERVISEUR',
  CAISSE_LOCATION: 'CAISSE_LOCATION'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Module = {
  PHARMACIE: 'PHARMACIE',
  RESTAURANT: 'RESTAURANT'
} as const

export type Module = (typeof Module)[keyof typeof Module]


export const ModePaiement = {
  CASH: 'CASH',
  CARTE: 'CARTE',
  MOBILE: 'MOBILE'
} as const

export type ModePaiement = (typeof ModePaiement)[keyof typeof ModePaiement]


export const Devise = {
  FRANC: 'FRANC',
  DOLLAR: 'DOLLAR'
} as const

export type Devise = (typeof Devise)[keyof typeof Devise]


export const StatutCommande = {
  EN_ATTENTE: 'EN_ATTENTE',
  EN_PREPARATION: 'EN_PREPARATION',
  SERVI: 'SERVI',
  PAYE: 'PAYE'
} as const

export type StatutCommande = (typeof StatutCommande)[keyof typeof StatutCommande]


export const StatutCommandeBar = {
  EN_COURS: 'EN_COURS',
  VALIDEE: 'VALIDEE',
  ANNULEE: 'ANNULEE'
} as const

export type StatutCommandeBar = (typeof StatutCommandeBar)[keyof typeof StatutCommandeBar]


export const TypeMouvementStock = {
  ENTREE: 'ENTREE',
  SORTIE: 'SORTIE'
} as const

export type TypeMouvementStock = (typeof TypeMouvementStock)[keyof typeof TypeMouvementStock]


export const TypeBien = {
  APPARTEMENT: 'APPARTEMENT',
  BUREAU: 'BUREAU',
  LOCAL_COMMERCIAL: 'LOCAL_COMMERCIAL',
  AUTRE: 'AUTRE'
} as const

export type TypeBien = (typeof TypeBien)[keyof typeof TypeBien]


export const NiveauBien = {
  REZ_DE_CHAUSSEE: 'REZ_DE_CHAUSSEE',
  N1: 'N1',
  N2: 'N2',
  N3: 'N3',
  N4: 'N4'
} as const

export type NiveauBien = (typeof NiveauBien)[keyof typeof NiveauBien]


export const EtatBien = {
  LIBRE: 'LIBRE',
  OCCUPE: 'OCCUPE',
  MAINTENANCE: 'MAINTENANCE'
} as const

export type EtatBien = (typeof EtatBien)[keyof typeof EtatBien]


export const StatutContrat = {
  ACTIF: 'ACTIF',
  TERMINE: 'TERMINE',
  EN_ATTENTE: 'EN_ATTENTE'
} as const

export type StatutContrat = (typeof StatutContrat)[keyof typeof StatutContrat]


export const RolePersonnel = {
  SERVEUR: 'SERVEUR',
  BARMAN: 'BARMAN'
} as const

export type RolePersonnel = (typeof RolePersonnel)[keyof typeof RolePersonnel]


export const TypeDocument = {
  PHOTO: 'PHOTO',
  CONTRAT: 'CONTRAT',
  AUTRE: 'AUTRE'
} as const

export type TypeDocument = (typeof TypeDocument)[keyof typeof TypeDocument]